refactor(chat): 리뷰 기반 파일 구조 수정, main 최신 코드 반영
