+++ src/routes/_main/_chat/route.tsx
@@ -1,0 +1,6 @@
+import { createFileRoute } from "@tanstack/react-router";
+import ChatPage from "./chat-content";
+
+export const Route = createFileRoute("/_main/chat")({
+  component: ChatPage,
+});
